fix: alle 18 Go-Services kompilieren + Frontend baut
Systematische Fehlerüberprüfung und Behebung aller Kompilierungsfehler:

- Fix: *logger.Logger pointer-to-interface → logger.Logger (63 Dateien)
- Fix: result.RowsAffected() multi-value context (8 Dateien)
- Fix: Import-Pfade github.com/rentflow → github.com/jeckersberger/rentflow
- Fix: cfg.DatabaseURL → cfg.ConnectionString() (18 Services)
- Fix: Repository-Typen *pgxpool.Pool/*sql.DB → *database.PostgresPool
- Fix: QueryContext/ExecContext → Query/Exec (PostgresPool API)
- Fix: Invoice TaxRate type (float64 → custom type mit Casts)
- Fix: Auth-Service net/http Package-Namenskonflikt (alias)
- Fix: Unused imports/variables in diversen Services
- Fix: tsconfig.json jsx: react-jsx hinzugefügt
- Frontend: terser als devDependency hinzugefügt
- README: Roadmap-Checkboxen für Phase 0-4 aktualisiert

Ergebnis: 18/18 Services kompilieren, Frontend baut (184 Module)

+++ services/invoice-service/internal/application/dto.go
@@ -127,7 +127,7 @@ func InvoiceToDTO(inv *domain.Invoice) *InvoiceDTO {
 			Unit:        item.Unit,
 			UnitPrice:   item.UnitPrice,
 			TotalPrice:  item.TotalPrice,
-			TaxRate:     item.TaxRate,
+			TaxRate:     float64(item.TaxRate),
 			EquipmentID: item.EquipmentID,
 		}
 	}
@@ -148,7 +148,7 @@ func InvoiceToDTO(inv *domain.Invoice) *InvoiceDTO {
 		ClientTaxID:   inv.ClientTaxID,
 		Items:         items,
 		SubTotal:      inv.SubTotal,
-		TaxRate:       inv.TaxRate,
+		TaxRate:       float64(inv.TaxRate),
 		TaxAmount:     inv.TaxAmount,
 		Total:         inv.Total,
 		Currency:      inv.Currency,
@@ -178,7 +178,7 @@ func QuoteToDTO(quote *domain.Quote) *QuoteDTO {
 			Unit:        item.Unit,
 			UnitPrice:   item.UnitPrice,
 			TotalPrice:  item.TotalPrice,
-			TaxRate:     item.TaxRate,
+			TaxRate:     float64(item.TaxRate),
 			EquipmentID: item.EquipmentID,
 		}
 	}
@@ -198,7 +198,7 @@ func QuoteToDTO(quote *domain.Quote) *QuoteDTO {
 		ClientEmail: quote.ClientEmail,
 		Items:       items,
 		SubTotal:    quote.SubTotal,
-		TaxRate:     quote.TaxRate,
+		TaxRate:     float64(quote.TaxRate),
 		TaxAmount:   quote.TaxAmount,
 		Total:       quote.Total,
 		Currency:    quote.Currency,
